pyproject.toml
[project]
name = "mai-hw2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.4.1",
    "hydra-core>=1.3.2",
    "ipykernel>=7.1.0",
    "omegaconf>=2.3.0",
    "peft>=0.17.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.9.0",
    "torchinfo>=1.8.0",
    "torchvision>=0.24.0",
    "transformers>=4.57.1",
]
